Mirror the `JsObjectType` API and implement `JsRevocableProxy`
